move props to type
